HOMER implementation, Double Bubble & WiM bug fixes
Homer
-Start implementing HOMER, it's fairly glitchy & needs some work done it
but the core algorithm is written.
Double Bubble
-Having some serious issues with scaling on the Double Bubble, it's all
done besides from picking the object from the menu using the bubble
cursor. I discovered doing this that there are some flaws in my initial
bubble cursor implementation which are relatively minor but I'm going to
have to go back & fix up or it could pose an issue with implementing it
in existing projects.
World In Miniature
-Just fixed a minor issue that was causing a null pointer exception