-Added a constants class, image-plane cont
-Added a new class (CONSTANTS) where we can put constant static
variables that will not be changed. For example the name of the left
controller is put into the constants class as a reference which is
called from all of the other scripts. Therefore if the left controller
name isn't set to the default SteamVR left controller name the user can
modify the value in the constants class rather than having to rename it
or modify all the C# scripts.